[refactor] :recycle: split services by provider (#276)
