+++ src/services/balanceService.ts
@@ -4,29 +4,20 @@ import NodeCache from 'node-cache';
 import { IToken } from '../models/tokenModel';
 import { Logger } from '../helpers/loggerHelper';
 import { SIGNING_KEY } from '../config/constants';
-import { setupERC20 } from './contractSetupService';
 import { getTokenAddress } from './blockchainService';
 import { IBlockchain } from '../models/blockchainModel';
+import { setupERC20 } from './web3/contractSetupService';
 import {
   CurrencyType,
   FiatQuoteType,
   TokenInfoType,
   BalanceInfoType,
   TokenBalanceType,
-  walletBalanceInfoType
-} from '../types/common';
+  WalletBalanceInfoType
+} from '../types/commonType';
 
 // Initialize the cache with a 5-minute TTL (Time To Live)
 const priceCache = new NodeCache({ stdTTL: 300, checkperiod: 320 });
-
-/**
- * API endpoints for fiat currency conversion rates
- */
-const API_URLs: [CurrencyType, string][] = [
-  ['UYU', 'https://criptoya.com/api/ripio/USDT/UYU'],
-  ['ARS', 'https://criptoya.com/api/ripio/USDT/ARS'],
-  ['BRL', 'https://criptoya.com/api/ripio/USDT/BRL']
-];
 
 /**
  * Fetches the balance of a specific token for a given address
@@ -35,7 +26,7 @@ const API_URLs: [CurrencyType, string][] = [
  * @param {string} address - Address to check balance for
  * @returns {Promise<string>} Token balance as a string
  */
-export async function getContractBalance(
+async function getContractBalance(
   contractAddress: string,
   signer: ethers.Wallet,
   address: string
@@ -58,30 +49,11 @@ export async function getContractBalance(
 }
 
 /**
- * Fetches fiat quotes from external APIs
- * @returns {Promise<FiatQuoteType[]>} Array of fiat currency quotes
- */
-export async function getFiatQuotes(): Promise<FiatQuoteType[]> {
-  return Promise.all(
-    API_URLs.map(async ([currency, url]) => {
-      try {
-        const response = await fetch(url);
-        const data = await response.json();
-        return { currency, rate: data.bid };
-      } catch (error) {
-        Logger.error('getFiatQuotes', `Error fetching ${currency} quote:`, error);
-        return { currency, rate: 1 }; // Fallback to 1:1 rate
-      }
-    })
-  );
-}
-
-/**
  * Fetches token prices from Binance API using USDT pairs
  * @param {string[]} symbols - Array of token symbols to fetch prices for
  * @returns {Promise<Map<string, number>>} Map of token symbols to their USD prices
  */
-export async function getTokenPrices(symbols: string[]): Promise<Map<string, number>> {
+async function getTokenPrices(symbols: string[]): Promise<Map<string, number>> {
   try {
     const priceMap = new Map<string, number>();
 
@@ -154,7 +126,7 @@ export async function getTokenPrices(symbols: string[]): Promise<Map<string, num
  * @param {number} chanId - Chain ID to filter tokens
  * @returns {Promise<TokenInfoType[]>} Array of tokens with current price information
  */
-export async function getTokenInfo(tokens: IToken[], chanId: number): Promise<TokenInfoType[]> {
+async function getTokenInfo(tokens: IToken[], chanId: number): Promise<TokenInfoType[]> {
   const chainTokens = tokens.filter((token) => token.chain_id === chanId);
   const symbols = [...new Set(chainTokens.map((token) => token.symbol))];
 
@@ -241,13 +213,13 @@ export function calculateBalancesTotals(balances: BalanceInfoType[]): Record<Cur
  * @param {ethers.Contract} tokenContract - Token contract instance
  * @param {string} walletAddress - Wallet address to check
  * @param {string} amountToCheck - Amount to check in wallet
- * @returns {Promise<walletBalanceInfoType>} Wallet balance information
+ * @returns {Promise<WalletBalanceInfoType>} Wallet balance information
  */
 export async function verifyWalletBalance(
   tokenContract: ethers.Contract,
   walletAddress: string,
   amountToCheck: string
-): Promise<walletBalanceInfoType> {
+): Promise<WalletBalanceInfoType> {
   const symbol: string = await tokenContract.symbol();
   Logger.log(
     'verifyWalletBalance',
@@ -263,7 +235,7 @@ export async function verifyWalletBalance(
     `Balance of wallet ${walletAddress}: ${walletBalanceFormatted} ${symbol}`
   );
 
-  const result: walletBalanceInfoType = {
+  const result: WalletBalanceInfoType = {
     walletBalance: walletBalanceFormatted,
     amountToCheck,
     enoughBalance: walletBalance.gte(amountToCheckFormatted)
@@ -278,14 +250,14 @@ export async function verifyWalletBalance(
  * @param {string} tokenContractAddress - Token contract address
  * @param {string} walletAddress - Wallet address to check
  * @param {string} amountToCheck - Amount to check in wallet
- * @returns {Promise<walletBalanceInfoType>} Wallet balance information
+ * @returns {Promise<WalletBalanceInfoType>} Wallet balance information
  */
 export async function verifyWalletBalanceByTokenAddress(
   blockchainConfig: IBlockchain,
   tokenContractAddress: string,
   walletAddress: string,
   amountToCheck: string
-): Promise<walletBalanceInfoType> {
+): Promise<WalletBalanceInfoType> {
   const provider = new ethers.providers.JsonRpcProvider(blockchainConfig.rpc);
   const backendSigner = new ethers.Wallet(SIGNING_KEY!, provider);
   const tokenContract: ethers.Contract = await setupERC20(tokenContractAddress, backendSigner);
@@ -299,7 +271,7 @@ export async function verifyWalletBalanceByTokenAddress(
  * @param {string} tokenSymbol - Symbol of the token
  * @param {string} walletAddress - Wallet address to check
  * @param {string} amountToCheck - Amount to check in wallet
- * @returns {Promise<walletBalanceInfoType>} Wallet balance information
+ * @returns {Promise<WalletBalanceInfoType>} Wallet balance information
  */
 export async function verifyWalletBalanceByTokenSymbol(
   blockchainConfig: IBlockchain,
@@ -307,7 +279,7 @@ export async function verifyWalletBalanceByTokenSymbol(
   tokenSymbol: string,
   walletAddress: string,
   amountToCheck: string
-): Promise<walletBalanceInfoType> {
+): Promise<WalletBalanceInfoType> {
   const provider = new ethers.providers.JsonRpcProvider(blockchainConfig.rpc);
   const backendSigner = new ethers.Wallet(SIGNING_KEY!, provider);
   const tokenContractAddress = getTokenAddress(blockchainConfig, blockchainTokens, tokenSymbol);
@@ -321,14 +293,14 @@ export async function verifyWalletBalanceByTokenSymbol(
  * @param {string} tokenAddress - Token contract address
  * @param {string} walletAddress - Wallet address to check
  * @param {string} amountToCheck - Amount to check in wallet
- * @returns {Promise<walletBalanceInfoType>} Wallet balance information
+ * @returns {Promise<WalletBalanceInfoType>} Wallet balance information
  */
 export async function verifyWalletBalanceInRpc(
   rpcUrl: string,
   tokenAddress: string,
   walletAddress: string,
   amountToCheck: string
-): Promise<walletBalanceInfoType> {
+): Promise<WalletBalanceInfoType> {
   const provider = new ethers.providers.JsonRpcProvider(rpcUrl);
   const erc20Abi = [
     'function transfer(address to, uint256 amount) returns (bool)',
