+++ src/services/blockchainService.ts
@@ -2,17 +2,17 @@ import { ethers } from 'ethers';
 
 import { IToken } from '../models/tokenModel';
 import { Logger } from '../helpers/loggerHelper';
-import { getEntryPointABI } from './bucketService';
-import { getBlockchain } from './mongo/mongoService';
+import { getEntryPointABI } from './gcp/gcpService';
 import { IBlockchain } from '../models/blockchainModel';
-import { setupContracts } from './contractSetupService';
+import { setupContracts } from './web3/contractSetupService';
 import { generatePrivateKey } from '../helpers/SecurityHelper';
-import { ensurePaymasterHasEnoughEth } from './paymasterService';
+import { ensurePaymasterHasEnoughEth } from './web3/paymasterService';
+import { mongoBlockchainService } from './mongo/mongoBlockchainService';
 import {
   TokenAddressesType,
-  setupContractReturnType,
+  SetupContractReturnType,
   CheckBalanceConditionsResultType
-} from '../types/common';
+} from '../types/commonType';
 import {
   USER_SIGNER_MIN_BALANCE,
   BACKEND_SIGNER_MIN_BALANCE,
@@ -187,14 +187,16 @@ export async function checkBlockchainConditions(
   fromNumber: string
 ): Promise<CheckBalanceConditionsResultType> {
   try {
-    const blockchain: IBlockchain | null = await getBlockchain(networkConfig.chain_id);
+    const blockchain: IBlockchain | null = await mongoBlockchainService.getBlockchain(
+      networkConfig.chain_id
+    );
 
     if (!blockchain) {
       throw new Error(`Blockchain with chain_id ${networkConfig.chain_id} not found`);
     }
 
     const privateKey = generatePrivateKey(fromNumber);
-    const setupContractsResult: setupContractReturnType = await setupContracts(
+    const setupContractsResult: SetupContractReturnType = await setupContracts(
       blockchain,
       privateKey,
       fromNumber
