+++ src/services/lifi/lifiService.ts
@@ -1,8 +1,8 @@
 import { Signer } from 'ethers-5';
 import { LiFi, RouteOptions, RoutesRequest } from '@lifi/sdk';
 
-import { Logger } from '../helpers/loggerHelper';
-import { LIFI_TYPE, LIFI_SLIPPAGE } from '../config/constants';
+import { Logger } from '../../helpers/loggerHelper';
+import { LIFI_TYPE, LIFI_SLIPPAGE } from '../../config/constants';
 
 /**
  * Performs a token swap from USDC to WETH using the LiFi SDK.
