+++ src/services/notificationService.ts
@@ -1,218 +1,27 @@
-import axios from 'axios';
-import { ethers } from 'ethers';
 import NodeCache from 'node-cache';
-import { channels as PushAPIChannels, payloads as PushAPIPayloads } from '@pushprotocol/restapi';
 
-import { getUser } from './mongo/mongoService';
 import { Logger } from '../helpers/loggerHelper';
-import { getNetworkConfig } from './networkService';
+import { pushService } from './push/pushService';
 import { IBlockchain } from '../models/blockchainModel';
-import { IUser, IUserWallet } from '../models/userModel';
-import { getTemplate, templateEnum } from './templateService';
+import { mongoUserService } from './mongo/mongoUserService';
+import { chatizaloService } from './chatizalo/chatizaloService';
 import { isValidPhoneNumber } from '../helpers/validationHelper';
+import { chatizaloOperatorReplyType } from '../types/chatizaloType';
+import { mongoBlockchainService } from './mongo/mongoBlockchainService';
+import { templateEnum, mongoTemplateService } from './mongo/mongoTemplateService';
+import {
+  BOT_DATA_TOKEN,
+  CHATTERPAY_NFTS_SHARE_URL,
+  NOTIFICATION_TEMPLATE_CACHE_TTL
+} from '../config/constants';
 import {
   LanguageEnum,
   ITemplateSchema,
   NotificationEnum,
   NotificationTemplatesTypes
 } from '../models/templateModel';
-import {
-  BOT_API_URL,
-  PUSH_ENABLED,
-  PUSH_NETWORK,
-  BOT_DATA_TOKEN,
-  PUSH_ENVIRONMENT,
-  DEFAULT_CHAIN_ID,
-  CHATTERPAY_DOMAIN,
-  PUSH_CHANNEL_ADDRESS,
-  GCP_CLOUD_TRACE_ENABLED,
-  PUSH_CHANNEL_PRIVATE_KEY,
-  BOT_NOTIFICATIONS_ENABLED,
-  CHATTERPAY_NFTS_SHARE_URL,
-  NOTIFICATION_TEMPLATE_CACHE_TTL,
-  SETTINGS_NOTIFICATION_LANGUAGE_DFAULT
-} from '../config/constants';
-
-interface OperatorReplyPayload {
-  data_token: string;
-  channel_user_id: string;
-  message: string;
-}
 
 const notificationTemplateCache = new NodeCache({ stdTTL: NOTIFICATION_TEMPLATE_CACHE_TTL });
-
-/**
- * Retrieves the wallet for a specific chain_id from a user's wallet array.
- * This function is internal to avoid circular imports between userService and notificationService.
- * @param {IUserWallet[]} wallets - The array of wallets to search through.
- * @param {number} chainId - The chain_id to filter the wallet.
- * @returns {IUserWallet | null} The wallet corresponding to the provided chain_id, or null if no matching wallet is found.
- */
-export const getUserWalletByChainIdInternal = (
-  wallets: IUserWallet[],
-  chainId: number
-): IUserWallet | null => {
-  const wallet = wallets.find((w) => w.chain_id === chainId);
-  return wallet || null;
-};
-
-/**
- * Sends an operator reply to the API.
- *
- * @param payload
- * @returns
- */
-async function sendBotNotification(
-  payload: OperatorReplyPayload,
-  traceHeader?: string
-): Promise<string> {
-  try {
-    if (!BOT_NOTIFICATIONS_ENABLED) {
-      Logger.info(
-        'sendBotNotification',
-        `Bot notifications are disabled. Omitted payload: ${JSON.stringify(payload)}`
-      );
-      return '';
-    }
-
-    const headers: { [key: string]: string } = {
-      'Content-Type': 'application/json'
-    };
-
-    if (GCP_CLOUD_TRACE_ENABLED && traceHeader) {
-      headers['X-Cloud-Trace-Context'] = traceHeader;
-    }
-
-    const sendMsgEndpint = `${BOT_API_URL}/chatbot/conversations/send-message`;
-    const response = await axios.post(sendMsgEndpint, payload, {
-      headers
-    });
-    Logger.log(
-      'sendBotNotification',
-      'API Response:',
-      payload.channel_user_id,
-      payload.message,
-      response.data
-    );
-    return response.data;
-  } catch (error) {
-    Logger.error('sendBotNotification', (error as Error).message);
-    throw error;
-  }
-}
-
-/**
- * Send Push Notificaiton
- *
- * @param title Notification Title
- * @param msg Notification Message
- * @param type 1 -> Broadcast, 3 -> Targeted
- * @param identityType // 0 -> Minimal, 2 -> Direct Payload
- */
-export async function sendPushNotificaton(
-  title: string,
-  msg: string,
-  channelUserId: string,
-  type: number = 3,
-  identityType: number = 2
-): Promise<boolean> {
-  try {
-    if (!PUSH_ENABLED) {
-      Logger.info('sendPushNotificaton', `Push notifications are disabled.`);
-      return true;
-    }
-
-    const user: IUser | null = await getUser(channelUserId);
-    if (!user) {
-      Logger.log(
-        'sendPushNotificaton',
-        `Push notification not sent: Invalid user in the database for phone number ${channelUserId}`
-      );
-      return false;
-    }
-
-    const userWallet: IUserWallet | null = getUserWalletByChainIdInternal(
-      user.wallets,
-      DEFAULT_CHAIN_ID
-    );
-    if (!userWallet) {
-      Logger.log(
-        'sendPushNotificaton',
-        `Push notification not sent: Invalid EOA Wallet in the database for phone number ${channelUserId}`
-      );
-      return false;
-    }
-
-    let { wallet_eoa } = userWallet;
-    wallet_eoa = wallet_eoa.startsWith('0x') ? wallet_eoa : `0x${wallet_eoa}`;
-
-    const signer = new ethers.Wallet(PUSH_CHANNEL_PRIVATE_KEY);
-    const apiResponse = await PushAPIPayloads.sendNotification({
-      signer,
-      type,
-      identityType,
-      notification: {
-        title,
-        body: msg
-      },
-      payload: {
-        title,
-        body: msg,
-        cta: CHATTERPAY_DOMAIN,
-        img: `${CHATTERPAY_DOMAIN}/assets/images/home/logo.png`
-      },
-      recipients: `eip155:${PUSH_NETWORK}:${wallet_eoa}`,
-      channel: `eip155:${PUSH_NETWORK}:${PUSH_CHANNEL_ADDRESS}`,
-      env: PUSH_ENVIRONMENT
-    });
-
-    Logger.log(
-      'sendPushNotificaton',
-      `Push notification sent successfully to ${channelUserId},  ${wallet_eoa}:`,
-      apiResponse.status,
-      apiResponse.statusText
-    );
-    return true;
-  } catch (error) {
-    Logger.error(
-      'sendPushNotificaton',
-      `Error sending Push Notification to ${channelUserId}:`,
-      error instanceof Error ? error.message : 'Unknown'
-    );
-    return false;
-  }
-}
-
-/**
- * Gets user language based on the phone number.
- *
- * @param phoneNumber
- * @returns
- */
-export const getUserSettingsLanguage = async (phoneNumber: string): Promise<LanguageEnum> => {
-  let language: LanguageEnum = SETTINGS_NOTIFICATION_LANGUAGE_DFAULT as LanguageEnum;
-  try {
-    const user: IUser | null = await getUser(phoneNumber);
-    if (user && user.settings) {
-      const userLanguage = user.settings.notifications.language;
-      if (Object.values(LanguageEnum).includes(userLanguage as LanguageEnum)) {
-        language = userLanguage as LanguageEnum;
-      } else {
-        Logger.warn(
-          'getUserSettingsLanguage',
-          `Invalid language detected for user ${phoneNumber}, defaulting to ${SETTINGS_NOTIFICATION_LANGUAGE_DFAULT}`
-        );
-      }
-    }
-  } catch (error: unknown) {
-    // avoid throw error
-    Logger.error(
-      'getUserSettingsLanguage',
-      `Error getting user settings language for ${phoneNumber}, error: ${(error as Error).message}`
-    );
-  }
-  return language;
-};
 
 /**
  * Get Notification Template based on channel User Id and Notification Type
@@ -233,10 +42,11 @@ async function getNotificationTemplate(
       return cachedTemplate as { title: string; message: string };
     }
 
-    const userLanguage: LanguageEnum = await getUserSettingsLanguage(channelUserId);
+    const userLanguage: LanguageEnum =
+      await mongoUserService.getUserSettingsLanguage(channelUserId);
 
     const notificationTemplates: NotificationTemplatesTypes | null =
-      await getTemplate<ITemplateSchema>(templateEnum.NOTIFICATIONS);
+      await mongoTemplateService.getTemplate<ITemplateSchema>(templateEnum.NOTIFICATIONS);
     if (!notificationTemplates) {
       Logger.warn('getNotificationTemplate', 'Notifications Templates not found');
       return defaultNotification;
@@ -273,66 +83,6 @@ async function getNotificationTemplate(
 }
 
 /**
- * Subscribe User To Push Channel
- *
- * @param user_private_key
- * @param user_address
- * @returns
- */
-export async function subscribeToPushChannel(
-  user_private_key: string,
-  user_address: string
-): Promise<boolean> {
-  try {
-    let userPrivateKeyFormatted = user_private_key;
-    let userAddressFormatted = user_address;
-
-    if (!user_private_key.startsWith('0x')) {
-      userPrivateKeyFormatted = `0x${user_private_key}`;
-    }
-    if (!user_address.startsWith('0x')) {
-      userAddressFormatted = `0x${user_address}`;
-    }
-
-    const signer = new ethers.Wallet(userPrivateKeyFormatted);
-    const subscriptionResponse = await PushAPIChannels.subscribe({
-      signer,
-      channelAddress: `eip155:${PUSH_NETWORK}:${PUSH_CHANNEL_ADDRESS}`,
-      userAddress: `eip155:${PUSH_NETWORK}:${userAddressFormatted}`,
-      onSuccess: () => {
-        Logger.log(
-          'subscribeToPushChannel',
-          `${userAddressFormatted} successfully subscribed to Push Protocol Channel`
-        );
-      },
-      onError: (error: unknown) => {
-        Logger.error(
-          'subscribeToPushChannel',
-          `Error trying to subscribe ${userAddressFormatted} to Push Protocol channel:`,
-          error
-        );
-      },
-      env: PUSH_ENVIRONMENT
-    });
-
-    Logger.log(
-      'subscribeToPushChannel',
-      `${userAddressFormatted} Push Protocol Subscription Response:`,
-      JSON.stringify(subscriptionResponse)
-    );
-    return true;
-  } catch (error) {
-    // Avoid throwing an error if subscribing to the push channel fails
-    Logger.error(
-      'subscribeToPushChannel',
-      `Error trying to subscribe ${user_address} to Push Channel:`,
-      error instanceof Error ? error.message : 'Unknown'
-    );
-    return false;
-  }
-}
-
-/**
  * Sends wallet creation notification.
  *
  * @param address_of_user
@@ -354,7 +104,7 @@ export async function sendWalletCreationNotification(
     );
     const formattedMessage = message.replace('[PREDICTED_WALLET_EOA_ADDRESS]', address_of_user);
 
-    sendPushNotificaton(title, formattedMessage, channel_user_id); // avoid await
+    pushService.sendPushNotificaton(title, formattedMessage, channel_user_id); // avoid await
   } catch (error) {
     Logger.error('sendWalletCreationNotification', error);
     throw error;
@@ -393,14 +143,14 @@ export async function sendTransferNotification(
       .replaceAll('[AMOUNT]', amount)
       .replaceAll('[TOKEN]', token);
 
-    const payload: OperatorReplyPayload = {
+    const payload: chatizaloOperatorReplyType = {
       data_token: BOT_DATA_TOKEN!,
       channel_user_id,
       message: formattedMessage
     };
 
-    const data = await sendBotNotification(payload, traceHeader);
-    sendPushNotificaton(title, formattedMessage, channel_user_id); // avoid await
+    const data = await chatizaloService.sendBotNotification(payload, traceHeader);
+    pushService.sendPushNotificaton(title, formattedMessage, channel_user_id); // avoid await
     return data;
   } catch (error) {
     Logger.error('sendTransferNotification', error);
@@ -430,7 +180,7 @@ export async function sendSwapNotification(
 ): Promise<unknown> {
   try {
     Logger.log('sendSwapNotification', 'Sending swap notification');
-    const networkConfig: IBlockchain = await getNetworkConfig();
+    const networkConfig: IBlockchain = await mongoBlockchainService.getNetworkConfig();
 
     const resultString: string = `${Math.round(parseFloat(result) * 1e4) / 1e4}`;
     const { title, message } = await getNotificationTemplate(
@@ -446,14 +196,14 @@ export async function sendSwapNotification(
       .replaceAll('[EXPLORER]', networkConfig.explorer)
       .replaceAll('[TRANSACTION_HASH]', transactionHash);
 
-    const payload: OperatorReplyPayload = {
+    const payload: chatizaloOperatorReplyType = {
       data_token: BOT_DATA_TOKEN!,
       channel_user_id,
       message: formattedMessage
     };
 
-    const data = await sendBotNotification(payload, traceHeader);
-    sendPushNotificaton(title, formattedMessage, channel_user_id); // avoid await
+    const data = await chatizaloService.sendBotNotification(payload, traceHeader);
+    pushService.sendPushNotificaton(title, formattedMessage, channel_user_id); // avoid await
     return data;
   } catch (error) {
     Logger.error('sendSwapNotification', error);
@@ -486,14 +236,14 @@ export async function sendMintNotification(
       .replaceAll('[ID]', id)
       .replaceAll('[NFTS_SHARE_URL]', CHATTERPAY_NFTS_SHARE_URL);
 
-    const payload: OperatorReplyPayload = {
+    const payload: chatizaloOperatorReplyType = {
       data_token: BOT_DATA_TOKEN!,
       channel_user_id,
       message: formattedMessage
     };
 
-    const data = await sendBotNotification(payload, traceHeader);
-    sendPushNotificaton(title, formattedMessage, channel_user_id); // avoid await
+    const data = await chatizaloService.sendBotNotification(payload, traceHeader);
+    pushService.sendPushNotificaton(title, formattedMessage, channel_user_id); // avoid await
     return data;
   } catch (error) {
     Logger.error('sendMintNotification', (error as Error).message);
@@ -524,7 +274,7 @@ export async function sendOutgoingTransferNotification(
     Logger.log('sendOutgoingTransferNotification', 'Sending outgoing transfer notification');
     if (!isValidPhoneNumber(channel_user_id)) return '';
 
-    const networkConfig: IBlockchain = await getNetworkConfig();
+    const networkConfig: IBlockchain = await mongoBlockchainService.getNetworkConfig();
 
     const { title, message } = await getNotificationTemplate(
       channel_user_id,
@@ -537,14 +287,14 @@ export async function sendOutgoingTransferNotification(
       .replaceAll('[EXPLORER]', networkConfig.explorer)
       .replaceAll('[TX_HASH]', txHash);
 
-    const payload: OperatorReplyPayload = {
+    const payload: chatizaloOperatorReplyType = {
       data_token: BOT_DATA_TOKEN!,
       channel_user_id,
       message: formattedMessage
     };
 
-    const data = await sendBotNotification(payload, traceHeader);
-    sendPushNotificaton(title, formattedMessage, channel_user_id); // avoid await
+    const data = await chatizaloService.sendBotNotification(payload, traceHeader);
+    pushService.sendPushNotificaton(title, formattedMessage, channel_user_id); // avoid await
     return data;
   } catch (error) {
     Logger.error('sendOutgoingTransferNotification', error);
@@ -574,14 +324,14 @@ export async function sendUserInsufficientBalanceNotification(
       NotificationEnum.user_balance_not_enough
     );
 
-    const payload: OperatorReplyPayload = {
+    const payload: chatizaloOperatorReplyType = {
       data_token: BOT_DATA_TOKEN!,
       channel_user_id,
       message
     };
 
-    const data = await sendBotNotification(payload, traceHeader);
-    sendPushNotificaton(title, message, channel_user_id); // avoid await
+    const data = await chatizaloService.sendBotNotification(payload, traceHeader);
+    pushService.sendPushNotificaton(title, message, channel_user_id); // avoid await
     return data;
   } catch (error) {
     Logger.error('sendUserInsufficientBalanceNotification', error);
@@ -611,14 +361,14 @@ export async function sendNoValidBlockchainConditionsNotification(
       NotificationEnum.no_valid_blockchain_conditions
     );
 
-    const payload: OperatorReplyPayload = {
+    const payload: chatizaloOperatorReplyType = {
       data_token: BOT_DATA_TOKEN!,
       channel_user_id,
       message
     };
 
-    const data = await sendBotNotification(payload, traceHeader);
-    sendPushNotificaton(title, message, channel_user_id); // avoid await
+    const data = await chatizaloService.sendBotNotification(payload, traceHeader);
+    pushService.sendPushNotificaton(title, message, channel_user_id); // avoid await
     return data;
   } catch (error) {
     Logger.error('sendNoValidBlockchainConditionsNotification', error);
@@ -648,14 +398,14 @@ export async function sendInternalErrorNotification(
       NotificationEnum.internal_error
     );
 
-    const payload: OperatorReplyPayload = {
+    const payload: chatizaloOperatorReplyType = {
       data_token: BOT_DATA_TOKEN!,
       channel_user_id,
       message
     };
 
-    const data = await sendBotNotification(payload, traceHeader);
-    sendPushNotificaton(title, message, channel_user_id); // avoid await
+    const data = await chatizaloService.sendBotNotification(payload, traceHeader);
+    pushService.sendPushNotificaton(title, message, channel_user_id); // avoid await
     return data;
   } catch (error) {
     Logger.error('sendInternalErrorNotification', error);
@@ -684,14 +434,14 @@ export async function SendConcurrecyOperationNotification(
       NotificationEnum.concurrent_operation
     );
 
-    const payload: OperatorReplyPayload = {
+    const payload: chatizaloOperatorReplyType = {
       data_token: BOT_DATA_TOKEN!,
       channel_user_id,
       message
     };
 
-    const data = await sendBotNotification(payload, traceHeader);
-    sendPushNotificaton(title, message, channel_user_id); // avoid await
+    const data = await chatizaloService.sendBotNotification(payload, traceHeader);
+    pushService.sendPushNotificaton(title, message, channel_user_id); // avoid await
     return data;
   } catch (error) {
     Logger.error('SendConcurrecyOperationNotification', error);
