+++ src/services/web3/simpleSwapService.ts
@@ -1,17 +1,17 @@
 import { ethers } from 'ethers';
 
-import { Logger } from '../helpers/loggerHelper';
+import { Logger } from '../../helpers/loggerHelper';
 import { setupERC20 } from './contractSetupService';
 import { addPaymasterData } from './paymasterService';
-import { IBlockchain } from '../models/blockchainModel';
+import { IBlockchain } from '../../models/blockchainModel';
 import { sendUserOperationToBundler } from './bundlerService';
 import { waitForUserOperationReceipt } from './userOpExecutorService';
 import { signUserOperation, createGenericUserOperation } from './userOperationService';
 import {
   TokenAddressesType,
   ExecuteSwapResultType,
-  setupContractReturnType
-} from '../types/common';
+  SetupContractReturnType
+} from '../../types/commonType';
 
 /**
  * Creates callData for token approval
@@ -131,7 +131,7 @@ async function executeOperation(
  */
 export async function executeSwap(
   networkConfig: IBlockchain,
-  setupContractsResult: setupContractReturnType,
+  setupContractsResult: SetupContractReturnType,
   entryPointContract: ethers.Contract,
   tokenAddresses: TokenAddressesType,
   amount: string
