+++ src/services/web3/userOperationService.ts
@@ -1,15 +1,15 @@
 import { ethers, BigNumber } from 'ethers';
 
-import { Logger } from '../helpers/loggerHelper';
-import { getUserOpHash } from '../helpers/userOperationHekper';
-import { PackedUserOperationType } from '../types/userOperation';
+import { Logger } from '../../helpers/loggerHelper';
+import { getUserOpHash } from '../../helpers/userOperationHekper';
+import { PackedUserOperationType } from '../../types/userOperationType';
 import {
   CALL_GAS_LIMIT,
   MAX_FEE_PER_GAS,
   PRE_VERIFICATION_GAS,
   VERIFICATION_GAS_LIMIT,
   MAX_PRIORITY_FEE_PER_GAS
-} from '../config/constants';
+} from '../../config/constants';
 
 /**
  * Creates a generic user operation for any type of transaction.
